refactor: processor 패키지 통합
