+++ src/main/java/com/bookspot/batch/step/StockStepConfig.java
@@ -4,8 +4,8 @@ import com.bookspot.batch.stock.data.CurrentLibrary;
 import com.bookspot.batch.stock.data.LibraryStock;
 import com.bookspot.batch.stock.data.LibraryStockCsvData;
 import com.bookspot.batch.stock.data.StockFileData;
-import com.bookspot.batch.stock.processor.LibraryStockProcessor;
-import com.bookspot.batch.stock.processor.file.StockFileInfoParser;
+import com.bookspot.batch.step.processor.csv.stock.LibraryStockProcessor;
+import com.bookspot.batch.step.processor.crawler.stock.StockFileInfoParser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.repository.JobRepository;
