+++ src/main/java/com/bookspot/batch/step/processor/csv/book/YearParser.java
@@ -1,11 +1,11 @@
-package com.bookspot.batch.book.processor;
+package com.bookspot.batch.step.processor.csv.book;
 
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-@Component
+@Service
 public class YearParser {
     private static final Pattern YEAR_PATTERN = Pattern.compile("\\d{4,}");
