+++ src/main/java/com/bookspot/batch/step/processor/csv/stock/IsbnValidationProcessor.java
@@ -1,5 +1,6 @@
-package com.bookspot.batch.stock.processor;
+package com.bookspot.batch.step.processor.csv.stock;
 
+import com.bookspot.batch.step.processor.csv.IsbnValidator;
 import com.bookspot.batch.stock.data.LibraryStockCsvData;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
