+++ src/main/java/com/bookspot/batch/stock/reader/StockReaderConfig.java
@@ -1,6 +1,6 @@
 package com.bookspot.batch.stock.reader;
 
-import com.bookspot.batch.book.processor.YearParser;
+import com.bookspot.batch.step.processor.csv.book.YearParser;
 import com.bookspot.batch.step.writer.file.stock.StockCsvMetadataCreator;
 import com.bookspot.batch.stock.data.LibraryStockCsvData;
 import lombok.RequiredArgsConstructor;
